third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-initialize-span-one-items-expected.txt

This test checks that content sized tracks are initialized using non-spanning grid items.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(gridAutoAndAuto, '').getPropertyValue('grid-template-columns') is "10px 90px"
PASS window.getComputedStyle(gridAutoAndAutoReversed, '').getPropertyValue('grid-template-columns') is "90px 10px"
PASS window.getComputedStyle(gridFixedMinContentFixedMaxContent, '').getPropertyValue('grid-template-columns') is "50px 50px"
PASS window.getComputedStyle(gridFixedMinContentFixedMaxContentTwoEmptyChildren, '').getPropertyValue('grid-template-columns') is "50px 50px"
PASS window.getComputedStyle(gridFixedMinContentFixedMaxContentFixedSizeChild, '').getPropertyValue('grid-template-columns') is "60px 40px"
PASS window.getComputedStyle(gridMaxContentAndAuto, '').getPropertyValue('grid-template-columns') is "80px 20px"
PASS window.getComputedStyle(gridMaxContentAndAutoReversed, '').getPropertyValue('grid-template-columns') is "90px 10px"
PASS window.getComputedStyle(gridMinContentAndAuto, '').getPropertyValue('grid-template-columns') is "10px 90px"
PASS window.getComputedStyle(gridMinContentAndAutoReversed, '').getPropertyValue('grid-template-columns') is "20px 80px"
PASS successfullyParsed is true

TEST COMPLETE

X
XXX XX XXX
X
XXX XX XXX
XXX XX XXX
XXX XX XXX
XXXXXX
XXX XX XXX
X
XXX XX XXX
X
XXX XX XXX
X
XXX XX XXX
X
XXX XX XXX
